Automating doc generation in build.sh

+++ build.sh
@@ -111,17 +111,23 @@ function clean()
 
 function doc()
 {
-    doxygen doc/Doxyfile
-		DATETODAY=$(date +%Y-%m-%d)
-		mkdir -p "../doc_temp_copy_$DATETODAY"
-		cp -r doc/html/* "../doc_temp_copy_$DATETODAY"
-		CURRENTBRANCH=$(git rev-parse --abbrev-ref HEAD)
-		git checkout gh-pages
-		cp -r "../doc_temp_copy_$DATETODAY/*" .
-		git commit -am "Documentation from $DATETODAY"
-		git push
-		git checkout "$CURRENTBRANCH"
-		rm -fr "../doc_temp_copy_$DATETODAY"
+		if [[ -z $(git status -s) ]]
+		then
+				doxygen doc/Doxyfile
+				DATETODAY=$(date +%Y-%m-%d)
+				mkdir -p "../doc_temp_copy_$DATETODAY"
+				cp -r doc/html/* "../doc_temp_copy_$DATETODAY"
+				CURRENTBRANCH=$(git rev-parse --abbrev-ref HEAD)
+				git checkout gh-pages
+				cp -r "../doc_temp_copy_$DATETODAY/*" .
+				git commit -am "Documentation from $DATETODAY"
+				git push
+				git checkout "$CURRENTBRANCH"
+				rm -fr "../doc_temp_copy_$DATETODAY"
+		else
+				echo -e "${RED}>>> You must commit your changes before running this command.${NC}\n"
+				git status
+		fi
 }
 
 function tests()
